Move use-slideshow to hooks folder
