+++ src/components/hooks/use-slideshow.js
